Camera

+++ WindowAPI/yaEntity.cpp
@@ -3,11 +3,11 @@
 namespace ya
 {
 	Entity::Entity()
-		: mID(reinterpret_cast<UINT64>(this))
+		: mID(reinterpret_cast<UINT32>(this))
 	{
 	}
 	Entity::Entity(const Entity& other)
-		: mID(reinterpret_cast<UINT64>(this))
+		: mID(reinterpret_cast<UINT32>(this))
 		, mName(other.mName)
 	{
 	}
